+++ updater/scripts/common.ps1
@@ -1,0 +1,11 @@
+
+function GetComparableVersion([string] $value)
+{
+    $value = $value -replace '^v', ''
+    try {
+        [System.Management.Automation.SemanticVersion]::Parse($value)
+    } catch {
+        Write-Warning "Failed to parse string '$value' as semantic version: $_"
+        $null
+    }
+}